fixed issue with string data properties
when using native FOAF properties should use $data_prop_prefix

There is an issue with one publication whose title is embraced by "" and this symbol is not allowed in xml.
Should remove publication or process titles before adding it.

Some titles also have a space in the beggining which is transformed to underscore by title routine.

Currently publishedAt property is valued to the publisher name, but the publisher type info (Magazine or Conference name) is missing.
Should add Magazine and Conferences as entities or should add a data property field to indicate it?